Moved .gitignore from robottests/ to /
It is possible to run Robot Framework anywhere in the project, so we need
to make sure that we are ignoring the generated files in all locations.

Before we only ignore the generated files inside the robottest/ folder.  By
moving .gitignore to root we can ignore the generated files in all locations.